Fixing errors wih db:migrate

+++ db/migrate/20151117132752_drop_volumestatuses_table.rb
@@ -1,5 +1,5 @@
 class DropVolumestatusesTable < ActiveRecord::Migration
   def change
-    drop_table :volumestatuses
+    drop_table :volumestatuses if ActiveRecord::Base.connection.table_exists? 'volumestatuses'
   end
 end
